Add an automated test for CTest Driver for TribitsExampleProject (#193)
This is the first fully automated test of TribitsCTestDriverCore.cmake that
actually runs ctest for everything, except the submit.  Actually, it can do a
submit to CDash too if asked (but it is off by default for obvious reasons).

To make this work, I added a hack for ignoring the extra repos file.  This
really needs to be cleared up better than this (but it works for now).

I also added printing of success doing the configure and build of each
package.  I added greps for a stronger test.

Next, I will add testing for failure cases and more printing and grepping to
pin down behavior.