Improvements to TribitsCTetsDriverCore.cmake and added testing (#193)
This is a major upgrade in the automated testing that is done with
TribitsCTestDriverCore.cmake.  This tests the major failure use cases to make
sure the CTest code behaves correctly.  As part of adding this automated
testing, I fixed a few defects and made other improvements:

* Improved the printout of the test failing indication logic.  (This is used
  in the new automated test.)

* Allow for override of <Project>_TRIBITS_DIR from env.  Before, this override
  was not possible.

WIP: Only run CTestDriver tests for CMake 3.3+ (#193)

For some reason, these tests do not work locally with CMake 2.8.11 but do work
with CMake 3.3.2.