Remove duplicate 'NAME' arg in CMAKE_PARSE_ARGUMENTS() (#181)
This generates a warning with CMAKE_PARSE_ARGUMENTS() but was ignored in
PARSE_ARGUMENTS().

This did not cause any TriBITS tests to fail but it did show a warning in the
test TriBITS_TestingFunctionMacro_UnitTests.  Howver, this was not noticed
before the last push.

This was not noticed until testing against Trilinos in trilinos/Trilinos#1378.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=243,notpassed=0 (0.58 min)
1) SERIAL_RELEASE => passed: passed=243,notpassed=0 (0.61 min)