Extend all-at-once mode to detect failing packages (#183)
This is now a fully functional all-at-once implementation of the CTest driver
that can detect failures to specific packages given the names of failing
tests.  It can map the names of failing tests for subpackages to the parent
package so it works just with top-level packages.  In the future, it might
make sense to only enable subpackackages to speed up the CI builds but for
now, I think this is the safest thing to do.

A few things worth mentioning:

* Generate all-at-once tests in two forms:
  - _AAO_: Without new CMake/CTest features (so it works with old versions of
    CMake/CTest) (TribitsExProj_DO_ALL_AT_ONCE empty or false)
  - _AAO__: With new CMake/CTest features (TribitsExProj_DO_ALL_AT_ONCE=TRUE)

* Set <Project>_ENABLE_ALL_PACKAGES=ON instead of each package enable.  This
  will scale better to larger projects with lots of packages, like CASL VERA.

* Added Python TribitsPackageTestNameUtils.py module and
  get-tribits-packages-from-last-tests-failed.py driver to help detect failing
  package given the names of failing tests.  (There are good unit tests for
  this as well).

* Added special handling of ALL_PACKAGES when failedPackages.txt file is read
  in.  This should ensure correct behavior in the all-at-once approach for a
  CI server.

* Added some whitespace to improve the printouts.

* Added todos for make dashboard target tests

* Added 'PBP_' to COMMON_ENV_ARGS var