Add all-at-once coverage and memory testing and tests (#183)
I added functionality for that all-at-once coverage and memory testing and
automated tests to make sure they work.

I also changed the CTestDriver CDash submit, memory, and coverage enable vars
to cache vars and use _CTEST_DRIVER_ at beginning.  This makes it easier to
find these options using ccmake, for example, or just in the cache file.

I also added greps for not running coverage and memory tests in the baisc AAO
case that does detailed grepping of the output.