Change name of var 'result' to 'eti_result' to avoid clash (#199)
Here I just changed the name of the 'result' var to 'eti_result' to avoid a
clash with the standard FindDoxygen.cmake module in CMake 3.9+.  Really we
should be initializing this var correctly but I am not sure how to do that
without studying the code more.  Therefore, I am not sure this a 100% fix for
this code (but it fixes the build error with Trilinos we were seeing reported
in #199).

This is a result of a latent bug in this code, strange cmake behavior for
variable scoping and functions, and a change in FindDoxygen.cmake.

I updated an existing test setting a dummy 'result' var to make sure that this
does not mess with the ETI work.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=246,notpassed=0 (0.61 min)
1) SERIAL_RELEASE => passed: passed=246,notpassed=0 (0.58 min)
2) MPI_DEBUG_CMAKE-3.6.2 => passed: passed=265,notpassed=0 (0.48 min)
3) SERIAL_RELEASE_CMAKE-3.6.2 => passed: passed=265,notpassed=0 (0.59 min)