Update for CTEST_LABELS_FOR_SUBPROJECTS (#183)
This replaces the usage of CTEST_ADD_SUBPROJECT() in the CTest -S script code
with just setting the var CTEST_LABELS_FOR_SUBPROJECTS in the CMakeLists.txt
file and the CTest -S driver script.  This submits to CDash correctly.

I added a strong automated test that verifies that LabelsForSubprojects is set
correctly in the generated DartConfiguration.tcl file.

This version works with the current labels-for-subprojects branch as of commit
bfbdbe1 "Help: Add notes for topic 'labels_for_subprojects'" on that branch.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=246,notpassed=0 (0.69 min)
1) SERIAL_RELEASE => passed: passed=246,notpassed=0 (0.58 min)
2) MPI_DEBUG_CMAKE-3.6.2 => passed: passed=265,notpassed=0 (0.48 min)
3) SERIAL_RELEASE_CMAKE-3.6.2 => passed: passed=265,notpassed=0 (0.59 min)