WIP: Add <Path> to CDashSubprojects.xml file (#183)
This adds the necessary code and data to set the <Path> element to submit to
CDash in hopes that it will result in CDash splitting out the coverage
results.

To make this work, I had to add a new attribute 'baseDirName' to the
<PackageDependencies> element along with the 'project' attribute and therefore
add it to the TribitsDependencies Python module.

I have not updated the automated tests so I suspect (and very much hope) that
some of them will fail.