Add <Path> to CDashSubprojectDependencies.xml file (#183)
This adds the necessary code and data to set the <Path> element to submit to
CDash in hopes that it will result in CDash splitting out the coverage
results.  Unfortunately, this did not trigger the behavior we had hoped, but
it is harmless to have this extra element in the XML input so there is no
reason not to keep this change.

To make this work, I had to add a new attribute 'baseDirName' to the
<PackageDependencies> element along with the 'project' attribute and therefore
add it to the TribitsDependencies Python module.  That gives what the
conversion function that creates the CDashSubprojectsDependencies.xml file.

I also updated the tests for updated XML files.  These automated tests really
pin down what is produced in the <Project>Dependencies.xml and
CDashSubprojectsDependencies.xml files.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=246,notpassed=0 (0.65 min)
1) SERIAL_RELEASE => passed: passed=246,notpassed=0 (0.63 min)
2) MPI_DEBUG_CMAKE-3.6.2 => passed: passed=265,notpassed=0 (0.60 min)
3) SERIAL_RELEASE_CMAKE-3.6.2 => passed: passed=265,notpassed=0 (0.59 min)