Run gitdist from anywhere in repo
Look at the current working directory when gitdist is executed,
determine the top-level git repo in that path, and change to it before
continuing.  This allows gitdist to be run from anywhere in a git
repository, including within a nesting of repositories.  For instance,
if you have a directory structure like the following
metaRepo/
  .git/
  .gitdist
  repo1/
    .git
  repo2/
    .git
  etc.
executing gitdist anywhere in that directory tree will change you to the
metaRepo directory, execute gitdist, and then leave you back where you
started when gitdist finishes.