Find the base repo looking for .gitdist[.default] (#203)
This is closing but does not quite work for all use cases.  It fails when
there is a set of nested git repos and .gitdist[.default] files like:

Trilinos/
  .git/
  .gitdist.default
  TriBITS/
    .git/
    .gitdist.default
    TriBITSDoc/
      .git/
  kokkos/
    .git/

When I run the new gitdist in Trilinos/ or Trilinos/kokkos/, it finds the
right directory to run in.  But when I try to run in Trilinos/TriBITS/ or
Trilinos/TriBITS/TriBITSDoc/, it always goes back to the base Trilinos/ dir.
It should stop at the first directory where .gitdist or .gitdist.default is
found.