Don't add tests when tests not enabled (#204)
This change makes is so that TRIBITS_ADD_[ADVANCED_]TEST() will not add a test
when ${PACKAGE_NAME}_ENABLE_TESTS and ${PARENT_PACKAGE_NAME}_ENABLE_TESTS are
not true.  A lot of people think this is the behavior anyway so let’s make it
the behavior anyway so we might as well make it the behavior. A more efficient
way to optionally enable tests is to put them in a test/ subdir and then
include that subdir with TRIBITS_ADD_TEST_SUBDIRECTORIES().

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=246,notpassed=0 (0.61 min)
1) SERIAL_RELEASE => passed: passed=246,notpassed=0 (0.65 min)
2) MPI_DEBUG_CMAKE-3.6.2 => passed: passed=265,notpassed=0 (0.81 min)
3) SERIAL_RELEASE_CMAKE-3.6.2 => passed: passed=265,notpassed=0 (0.78 min)