Colorize git commands run through gitdist
When opening a subprocess and passing the stdout to PIPE, the output is
buffered, losing the colors otherwise seen when running the command from
the command line.  One solution to keeping the colors is to redirect the
stdout and stderr of the subprocess to the stdout and stderr of the main
process.  Calling communicate() then makes the call block until the end
of the subprocess, otherwise gitdist would continue while the previous
git command was finishing, leading to interspersed lines of output.