Fix several issues with CONTRIBUTING.md
I think this is ready to use.

I did several things here:

Fixed formatting of nested bullets: You have to line up the '*' in the bullet
with the second letter of the above list text and there can't be any spaces
between the list items.

Make links to internal sections relative: You have to do this now that this
file is version controlled.  But I tested and you can link all over the place
and it works.