Clean up TribitsHelloWorld some, allow TPL_ENABLE_MPI to be undefiend (#190)
I cleaned up TribitsHelloWorld/CMakeLists.txt to make it shorter and compare
better against RawHelloWorld/CMakeLists.txt.  I removed a lot of the comments.
These comments are better made in the more general documentation.

Hopefully this makes it easier to create simple starter TriBITS projects by.
not having so specify MPI as a TPL or hack setting TPL_ENABLE_MPI off.

I updated the test of TribitsHelloWorld to not set TPL_ENABLE_MPI and updated
TriBITS to set it to FALSE when it is not set already.