Remove minimum CUDA version 3.2 (#222)
This was put in a *very* long time ago and 3.2 is way to low for any software
currently using CUDA and TriBITS (e.g. Tpetra currently requres version 7.5 or
above).

It is now expected that a user will install and point to a new-enough version
of CUDA and the packages that use CUDA will be required to assert that the
version is new enough and if not error out.