Add MESSAGE(FATAL_ERROR ...) to abort before tests run again (#183)
There is a strange defect in ctest -S or something that is resulting in the
tests being run again right before the ctest -S script
experimental_build_test.cmake exists.  Calling MESSAGE(FATAL_ERRROR ...) is
the only way that I know to abort the ctest -S script.  But since this is just
for the 'dashbaord' target and since the return value from that ctest -S
command is ignored, this is not really a problem (other than looking ugly).

NOTE: This problem did not seem to occur when directly running ctest -S
<some_scirpt> using TribitsCTestDriverCore.cmake.