Add TRIBITS_TPL type to read in <Package>Config.cmake to be read as TPL (ATDV-84, #63)
This is a fairly small change to the processing of TPLs that works for
pre-installed TriBITS packages.  Basically you just mark the TPL as
"TRIBITS_TPL" in the TPLsList.cmake file instead of pointing to a directory or
FindTPL<tplName>.cmake module file and then find_package(<tplName> REQUIRED)
is used to get the parts of the TPL.

At this point there are no automated tests to support and protect this.  This
is just a little refactoring (about 10 lines of code is all) that allowed us
to have a prototype TriBITS build system for an APP (SPARC) treat
pre-installed TriBITS packages as TPLs (see ATDV-84).  I really don't want to
even docuemnt this because I view this as just a small stepping-stone on the
path merging the concepts and implementation of TriBITS Packages and TPLs.
Therefore, I don't want to support this new TRIBITS_TPL TPL for very long and
only for this one customer.  I hope to be able to do enough refactoring of
TriBITS so that an APP can just list these pre-installed TriBITS packages as
packages in their Dependencies.cmake file and then have TriBITS automatically
treat them as TPLs and look for them with find_package().

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=260,notpassed=0 (0.70 min)
1) SERIAL_RELEASE => passed: passed=260,notpassed=0 (0.58 min)
2) MPI_DEBUG_CMAKE-3.6.2 => passed: passed=281,notpassed=0 (0.68 min)
3) SERIAL_RELEASE_CMAKE-3.6.2 => passed: passed=281,notpassed=0 (0.84 min)
Other local commits for this build/test group: c25a7df, a8a1f9f