Modify install-cmake.py to build CMake with SSL (#220)
For CMake to be able to submit to a drop site via HTTPS
the cmake/ctest executables must have been built with OpenSSL
support. This commit modifies the install-cmake.py script to:
  * Check if CMake is installed on the system
  * Configure CMake with CMake and OpenSSL support

The CMakeLists.txt for configuring CMake with CMake with is smart
enough to allow you to still build CMake even if you request OpenSSL
and it is not installed. The only requirement is that CMake be
configured with CMake instead of autotools. If CMake is not available
then install-cmake.py defaults to its previous behavior.