Add new TAAT() TEST_<IDX> COPY_FILES_TO_TEST_DIR block (#228)
This new feature makes it easy to copy a bunch of files to a newly created
TRIBITS_ADD_ADVANCED_TEST() working directory.

As part of this, I refactored the DriveAdvancedTest.cmake code to clean it up
and make it (hopefully) easier to understand.

A major part of this is that I added a new way to test TAAT() by running
TAAT() in a nested CMake project so that TAAT() can actually check that TAAT()
is printing the right things and behaving as it should.  This will be very
useful when more refactorings and performed and new feature are added to
TAAT().