Allow customization of number submit return attempts and delay and set new defaults (#230)
TriBITS was hardcoding the number of submit retry attempts to 25 and the delay
between attempts as 120 second.  That means that if there is a MySQL error
(which is happening right now with attempted submits of SPARC build data as
part of ATDV-84), then it will take 50 minutes to trying to submit before
moving on!