gitdist:  Add Default Development Branch... (#235)
Here's an initial implementation that allows a user to add default
development branches to the .gitdist file in the form

. master
otherRepo develop
thirdRepo master

If those default development branches aren't specified, master is
assumed.  The user can then pass _DEFAULT_BRANCH_ to any command that
accepts a reference and gitdist should replace _DEFAULT_BRANCH_ with the
appropriate branch from the .gitdist file.  The documentation for this
feature can be accessed with --dist-help=default-branch.