Update to work with NetCDF-4.5.1 or later
As of NetCDF-4.5.1 or later, it is no longer necessary to modify `NC_MAX_DIMS` and `NC_MAX_VARS` as the limits are no longer checked inernally in the library.  This PR modifies the module to determine the NetCDF version and if 4.5.1 or later, does not check the `NC_MAX_DIMS` or `NC_MAX_VARS` settings and issue warning if not modified.