Add check for package-with-subpackage commands in packages without subpackages (#200)
The SCOREC packages was generating strange configure errors when it was trying
to call commands for a package with subpackages but did not have any
subpackages.  The error message was terrible in this case.  I added a new
check and new test case to produce a good error message.  Hopefully that would
have made it easier to fix the error in the SCOREC package.

I also improved the error message to mention that you can call
TRIBITS_PACKAGE_DECL() and TRIBITS_PACKAGE_DEF() or TRIBITS_PACKAGE() before
TRIBITS_PACKAGE_POSTPROCESS().  That hopefully is a little better error
message?  In any case, I think we need to go over these error messages a
little more to make sure they provide the best info to the developer.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=263,notpassed=0 (0.72 min)
1) SERIAL_RELEASE => passed: passed=263,notpassed=0 (0.59 min)
2) MPI_DEBUG_CMAKE-3.6.2 => passed: passed=284,notpassed=0 (0.56 min)
3) SERIAL_RELEASE_CMAKE-3.6.2 => passed: passed=284,notpassed=0 (0.54 min)