Define <Project>_C_Standard as cache var and document (#239)
There is no C code in TribitsExampleProject so I could not set up a direct
test for the correct setting of the cache var ${PROJECT_NAME}_C_Standard.  But
since this is just a small amount of code I tested it with a quick build of
Zoltan under Trilinos (which has *.c files).

I added documentation to the Tribits Developers Guide since it is mostly
project developers who will want to change the C standard.  I did not directly
document this new option in the TriBITS Project Build Reference since I did
not think it needed to be documented there.  But since this is a cache var,
users can find this by examining the cache vars and the documentation for
those.  Hopefully that is enough.

ToDo: We need to add an example/test package that has C code so that we can
test C handling of real source code by TriBITS.