Set Travis CI test timeout to 60s (#240)
The Travis CI build did not have any timeout set for the tests.  Therefore, if
any one test hung, it would result in a timeout of the overall Travis CI build
and no test results would show up on the TriBITS CDash site.

This change gives what I think is a pretty reasonble default of 60s.