Add clean of ignored files in orig-dir (#242)
The script snapshot-dir.py now, by default, cleans out git ignored files in
the orig-dir/.  This is done so that files that may only have local git
ignores (i.e. in .git/info/excludes) in the orig-dir don't get copied into and
commited in the dest-dir.  This will resolve issues where a temp ignored file
in the orig-dir is accidentally copied and committed to the dest-dir repo as
reported in #242.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=263,notpassed=0 (0.72 min)
1) SERIAL_RELEASE => passed: passed=263,notpassed=0 (0.61 min)