checkin-test.py: Small --log-file fixups and a little more testing
* Moved --log-file into a more local place in the listing of options

* Added the echo of --log-file in the list of options in the log file

* Added test to check for default value for --log-file

Note that I did not add a new test just for the --log-file option but I did
add setting to to an existing test.  If you comment out setting --log-file to
a non-default file name that is passed in like
--log-file=checkin-test.something.out, then it will fail the automated test
that I put in.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=268,notpassed=0 (0.74 min)
1) SERIAL_RELEASE => passed: passed=268,notpassed=0 (0.61 min)
2) MPI_DEBUG_CMAKE-3.6.2 => passed: passed=289,notpassed=0 (0.56 min)
3) SERIAL_RELEASE_CMAKE-3.6.2 => passed: passed=289,notpassed=0 (0.55 min)