Add DEFAULT vars for a few vars use in Trilinos/ProjectName.cmake (trilinos/Trilinos#2464)
This defines and uses the defaults:

* ${PROJECT_NAME}_ENABLE_SECONDARY_TESTED_CODE_DEFAULT
* ${PROJECT_NAME}_GENERATE_REPO_VERSION_FILE_DEFAULT

This allows these vars to be set in the <projectDir>/ProjectName.cmake file
but then allow SET( ... CACHE ...) statements in include *.cmake files read in
through ${PROJECT_NAME}_CONFIGURE_OPTIONS_FILE to define the defaults for
these cache vars.

This is the idiom that should be used for all of these types of cache vars
with defaults set by TriBITS.

The default ${PROJECT_NAME}_ENABLE_SECONDARY_TESTED_CODE_DEFAULT is needed for
trilinos/Trilinos#2464 but I added the other default
${PROJECT_NAME}_GENERATE_REPO_VERSION_FILE_DEFAULT just because
Trilinos/ProjectName.cmake sets this as well.

I had to remove an anchor set in ${PROJECT_NAME}_ENABLE_SECONDARY_TESTED_CODE
in the file TribitsCTestDriverCore.cmake since it is now defined in the main
document with this change.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=268,notpassed=0 (0.71 min)
1) SERIAL_RELEASE => passed: passed=268,notpassed=0 (0.62 min)
2) MPI_DEBUG_CMAKE-3.6.2 => passed: passed=289,notpassed=0 (0.63 min)
3) SERIAL_RELEASE_CMAKE-3.6.2 => passed: passed=289,notpassed=0 (0.56 min)