Implement multi-phase runs of ctest -S driver (#259, TRIL-204)
This allows you to run ctest -S <script>.cmake multiple times doing different
steps in each in progression.  This is needed so that we can do start,
configure, and build on compile nodes of an HPC cluster and then run tests
only on the compute nodes.  This is needed for the ATDM Trilinos builds on the
SNL HPC machines (TRIL-171, TRIL-204).

This works for both the package-by-package mode and the all-at-once-mode.  And
it submits to and displays on CDash correctly in both modes.  That is quite
extraordinary for the package-by-pacakge mode.  This submits correctly to the
same CDash build by calling ctest_start(APPEND) on latter ctest -S invocations
so results go to the same CDash build.

I tested this locally and did trial submits and carefully inspected the
results on CDash (a newer version that can process the new all-at-once
format).  This looks solid.

Note that this was a pretty major refactoring of the logic for the
package-by-package mode and it is a lot cleaner, in my opinion.

I also improved some documentation about the 'dashboard' target not just
related to multi-invocation usage of TRIBITS_CTEST_DRIVER().