Set <Project>_CTEST_USE_NEW_AAO_FEATURES=TRUE as constant for CMake 3.10+ (#183)
There is no reason not to turn on new all-at-once features of CTest when a
CMake version 3.10 is used.  The default for
${PROJECT_NAME}_CTEST_DO_ALL_AT_ONCE is still FALSE and should be for some
time.  It is not until you set ${PROJECT_NAME}_CTEST_DO_ALL_AT_ONCE=TRUE that
you see any differences in behavior.  All mention of
${PROJECT_NAME}_CTEST_USE_NEW_AAO_FEATURES is removed from the documentation
but that var is kept as a convenient constatn.

As part of this I also set a proper cache var
${PROJECT_NAME}_CTEST_DO_ALL_AT_ONCE with default
${PROJECT_NAME}_CTEST_DO_ALL_AT_ONCE_DEFAULT.  I also updated the
documentation to note when this default should be set to TRUE.