Small cleanup and refactoring to get ready for project specialization (trilinos/Trilinos#3133)
I removed some old references to "trilinos" and replaced them with "project".

This is getting ready to add functionality for a TriBITS Project to specalize
logic for if a global rebuild is needed for certain files and directories.  I
added an object for that logic that will be easy to pass around.