Respond to env var MOCKPROGRAM_INOUT_FILE_OVERRIDE (trilnos/Trilinos#3133)
This feature needed to be added in order to support unit testing of the new
Trilinos get-changed-trilinos-packages.sh script (see trilinos/Trilinos#3133).
To unit test that script, you have to cd and run in the base Trilinos reppo
source tree.  But you don't want to be trying to write a
.mockprogram_inout.txt file in the real Trilinos source tree.  There, I needed
the ability to get mockprogram.py to read its input/output file from some
other location.