Some starting cleanup tutorial (#190)
Since this tutorial is in the tribits/doc/ directory, calling the
'tribits/doc/tribits_tutorial' is redundant.  Therefore, I changed the name to
'tribits/doc/tutorial'.

The driver script was trying to copy files to cupojoe.srn.sandia.gov.  Also,
it was removing output files automatically.  How can you debug problems if you
remove the output files?

Added README.md file to state that this is still in develpment and may not
build (I can't build it on my Linux machine).

This will need some refactoring to make this cleaner, more robust, and easier
to work with.