Make tests robust to CTEST_TEST_TYPE being set in env
When the TriBITS tests where run recently under CASL VERA, a bunch of these
failed because they had set CTEST_TEST_TYPE=Nightly in the env and that
interacted with some of these TriBITS tests badly.  I was able to reproduce
the test failures bu just running:

   env CTEST_TEST_TYPE=Nightly ctest -j16

After setting back to default CTEST_TEST_TYPE= in the env, all of these tests
passed.

NOTE: I did not add any other tests for this case because I did not think it
was worth the effort and the risk is very low.  If some env var interacts with
these tests and casues them to fail, it is easy enough to debug what is
happening and then reset the offending env var (like I did here) and then get
the tests to pass again.  Where there is more risk is when you can set an env
var and make tests pass that should otherwise fail (but I don't see how that
can happen given the nature of these very strong tests).