Remove some conditional code based on CMake versions less than 3.10.0 (trilinos/Trilinos#1761)
I took out logic to set the policy based on the CMake version of 3.3.  That
might change behavior of TriBITS projects so we will need to watch that.  If
that is a problem, we might just set the behavior to CMake 3.10.0 to be safe.
That might make the behavior of TriBITS less variable with new CMake versions.

I left in setting the policy CMP0054 which should make if() statements behave
better w.r.t. to varaible and constant interpretations.