WIP: Add proper HTML header and body and fix font sizes (trilinos/Trilinos#2933)
This increases the font size of the paragraphs at the top.  The are a lot
easier to read now.

To fix this, I had to define a proper <header> and <body>.  I carefully
inspected the generated HTML and it looks good both in the file that was
written and in the generated HTML email.  (But Outlook modifies the HTML a
good bit but my font changes seem to impact the email HTML correctly.)

Soon we will have this under some automated testing so we can protect this.