WIP: Add some more color for Passed, Failed, Not Run and Missing data
This should amke it easier to interpet the data in the various tables.

NOTE: We can work on a colorblind palette later.  I used functions
cdashColorPassed(), cdashColorFailed(), cdashColorNotRun() and
cdashColorMissing() to avoid hard-coding colors and make them easier to tweak
later and to put in a colorblide palette.

NOTE: I also put in a check that the top test history dict for 'Failed' and
'Not Run' tests matches the status of the nonpassing test dict itself.  This
helped me fix the test data for the automated tests.  But I did not put in a
unit test for this exception (but I should at some point).  This should also
help catch cases where there is a mismatch involving the 'date=YYYY-MM-DD'
field and the buildstarttime field.