Add support for CMAKE_DO_INSTALL (trilinos/Trilinos#2689)
I added tests for passing and failing installs.  I only implemented this for
the all-at-once approach since that is all we need right now for ATDM
Trilinos.  But we could implement it for the package-by-package mode if needed
without much trouble.

Note that as part of this I fixed an oversight where build failures that did
not cause test failures would be ignored in that all-at-once approach.  That
was wrong.  Now if there are any build or install failures, it will assume
that any (read that 'all') of the tested packages may have failed.  See the
long "NOTE" comment about this.  Long-story-short, this will only really
impact CI builds where there are build failures.  If there are just test
failures (the more common case), then only the packages with failing tests are
flagged as failed and will be enabled on the next CI iteration.