Don't add SOVERSION to shared libs if no Version.cmake file (#270)
Without this change, it was creating libs of the name
'lib<libname>.so.SOVERSION' (where 'SOVERSION' is literal!).  I added a test
that makes sure that a 'lib<libname>.so.SOVERSION' library is not created when
the Version.cmake files does not exist.

I also extended an existing test that check that the correct SOVERSION is
added to shared libs when Version.cmake exists.