WIP: Add list_cdash_notes_files_in_xml_file.sh (#278)
Can't actaully use this in checking the contents of the Notes.xml file because
that file does not get generated unless ctest_submit() is called and we can't
do CDash submits by default for lots of obvious reasons.

I am going to set up some tests of tribits_ctest_driver() that actually do
clones and updates and call ctest_update() that will then use this file.