WIP: Don't assume extra repos start on a tracking branch (#278)
The old code assumed that the current branch in an extra repo was on a tacking
branch.  That is a bad assumption.  The updated code just uses '@{u}' as the
tracking branch.

This was a real defect that was found through careful automated testing with
real git commands.  This is why we need some tests with real git commands to
validate our implementation.