Update logic for duplicate tests to allow for small diff in test 'time' (trilinos/Trilinos#2933)
For the testing day 2019-03-20 the SPARC Trilinos Integration script crashed
with the error:

  listOfDicts[1004]['time'] = '0.22' != listOfDicts[1002]['time'] = '0.23'

I reported this to Kitware but I don't know if they will fix it.  Therefore,
to make the tool robust for crasy behavor like this, I added code and unit
tests to allow a 0.1 relative error in the test 'time' to be considered the
same test.

This allows the SPARC Trilinos Integration script to run correctly on testing
day 2019-03-20.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=342,notpassed=0 (1.32 min)
1) SERIAL_RELEASE => passed: passed=342,notpassed=0 (1.60 min)