Change allowed rel error from 0.1 to 1.0 (#2933)
The SPARC Trilinos Integration run for 2019-03-22 crashed with the error:

  listOfDicts[1006]['time'] = '0.39' != listOfDicts[1003]['time'] = '0.28'

The account for this I changed the rel error calcuation to better compute
relative error (change const in denominator from 1.0 to 13-5) and allowed a
max rel error of 1.0.

If this fails again, I will make this an ajustable parameter that the user can
set and provide a better error message.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=342,notpassed=0 (1.12 min)
1) SERIAL_RELEASE => passed: passed=342,notpassed=0 (1.22 min)