Find git at global level for reuse (trilinos/Trilinos#4532)
There is a strong unit test for the generation of the <Project>RepoVersion.txt
file that ensures that git is found correctly and can be used.

NOTE: We switched from find_program(git) to find_package(Git) so that we could
use the GIT_VERSION_STRING later in the implementation of the
<Project>_version_date.h file (which will require git version 2.10+ in order
to work).