Fix test that was failing in 'dashboard' target
For some reason, when run with the 'dashbaord' target, the object file build
was not verbose when using 'make simplecxx' but is when just using 'make'.
Can't explain it.  But just doing 'make' results in the verbose commandline
for building the object file and allowing the grep to pass.

Now all 350 tests pass using 'make dashboard'!