Add vars for ninja job pool limits and always enable ninja testing if can be found (#281)
I made it so that the "Ninja" tests are always enabled and run if the 'ninja'
program can be found, even if the generator used for the TriBITS project
itself is using "Unix Makefiles".

I updated one of the Ninja tests to check that the right job pool vars are
being set in the CMake-generated rules.ninja file.  These are pretty solid
tests.

I also tested this manually with Trilinos and they seem to be working as they
should.