NETCDF: Fix search of netcdf_meta
To determine whether NetCDF is compiled for parallel, we search the `netcdf_meta.h` file for the setting of `NETCDF_PARALLEL`. The problem is that the NetCDF team recently added another setting which is named `NETCDF_PARALLEL4`.  We want the `NETCDF_PARALLEL` so to make sure we get that one only, we add a space to the end of the symbol `NETCDF_PARALLEL `