Fix documentation for <Project>_ENABLE_INSTALL_CMAKE_CONFIG_FILES default (trilinos/Trilinos#2698)
Turns out the default for <Project>_ENABLE_INSTALL_CMAKE_CONFIG_FILES is OFF,
not ON.  That was very confusing.

It is important that we test installs of TribitsExampleProject where there are
install failures and we ensure that the file <Project>Config.cmake gets
installed correctly and is usable when an installation fails.