Make usage of <Package>Config.cmake robust when there are broken packages (trilinos/Trilinos#2689)
Now when you use the 'install_package_by_package' target to install when there
are broken packages (that don't install everything), the installed
<Project>Config.cmake file can still be used as long as the user reqeusts only
packages that are not broken through the COMPONENTS feature.

This will allow SPARC to use an installation of Trilinos that might have
broken packages like Intrepid2 or Panzer that SPARC does not directly or
indirectly use.  This should allow a single build and install of Trilinos to
be robust for multiple customers so that any package build failures will ony
impact customers that actually use those packages.