gitdist:  Create Table of SHA1s (#286)
Add a new special command to gitdist (dist-repo-versions-table) such
that it can create a Markdown-formatted table of repositories and
corresponding commit information for easy inclusion in an issue tracking
system.  Provide a --dist-short option to limit the table to repository
name and corresponding SHA1 only.  Include appropriate testing and
documentation.