Add tribits/core install_package_by_package tests and <Project>_INSTALL_PBP_RUNNER option (#2689, ATDV-156)
Nice to have a test of the new install_package_by_package target (see #2689)
in the tribits/core test suite.  Also, also includes th new option
<Project>_INSTALL_PBP_RUNNER option needed for protected installs of Trilinos
using the 'jenkins' entity account (see ATDV-156).  This is such an obscure
use case I don't want to even try to explain this in the documentation.  It is
enough to have an automated test to protect this for now.