Add TRIBITS_REPORT_INVALID_TRIBITS_USAGE( and <Project>_ASSERT_CORRECT_TRIBITS_USAGE (#200)
This will allow CASL VERA to set:

  SET(${PROJECT_NAME}_ASSERT_CORRECT_TRIBITS_USAGE_DEFAULT WARNING)

so that it can upgrade TriBITS and then deal with the invalid usage warnings
after that.

This also sets ${PROJECT_NAME}_ENABLE_DEVELOPMENT_MODE_DEFAULT=ON by default
if it is not set by the project.  This is what it should have always been from
the beginning.  Hopefully this will not break any existing TriBITS projects
(but it is easy to fix by just setting the default in the
<projectDir>/ProjectName.cmake file).